+++ Sources/DockApp/DockApp.swift
@@ -21,46 +21,46 @@ let items: [DockItem] = [
 
 // MARK: - Constants
 
-let baseIconSize: Float = 48
-let maxScale: Float = 2.0
-let influenceRadius: Float = 150
-let iconPadding: Float = 8
-let dockHeight: Float = 64
+let baseIconSize: CGFloat = 48
+let maxScale: CGFloat = 2.0
+let influenceRadius: CGFloat = 150
+let iconPadding: CGFloat = 8
+let dockHeight: CGFloat = 64
 
 // MARK: - State
 
 class DockState {
-    var mouseX: Float = 0
-    var mouseY: Float = 0
+    var mouseX: CGFloat = 0
+    var mouseY: CGFloat = 0
     var minimizedAppIds: [String] = []
 }
 
 // MARK: - Magnification & Hit Testing
 
-func magnifiedSizes(mouseX: Float, iconCount: Int, screenWidth: Float) -> [Float] {
-    let totalBaseWidth = Float(iconCount) * baseIconSize + Float(iconCount - 1) * iconPadding
+func magnifiedSizes(mouseX: CGFloat, iconCount: Int, screenWidth: CGFloat) -> [CGFloat] {
+    let totalBaseWidth = CGFloat(iconCount) * baseIconSize + CGFloat(iconCount - 1) * iconPadding
     let startX = (screenWidth - totalBaseWidth) / 2
 
     return (0..<iconCount).map { i in
-        let iconCenterX = startX + Float(i) * (baseIconSize + iconPadding) + baseIconSize / 2
+        let iconCenterX = startX + CGFloat(i) * (baseIconSize + iconPadding) + baseIconSize / 2
         let distance = abs(mouseX - iconCenterX)
         if distance > influenceRadius { return baseIconSize }
         let t = 1.0 - (distance / influenceRadius)
-        let scale = 1.0 + (maxScale - 1.0) * (1.0 - cosf(t * .pi)) / 2.0
+        let scale = 1.0 + (maxScale - 1.0) * (1.0 - cos(t * .pi)) / 2.0
         return baseIconSize * scale
     }
 }
 
-func hoveredIconIndex(mouseX: Float, mouseY: Float, iconCount: Int, screenWidth: Float, screenHeight: Float) -> Int? {
+func hoveredIconIndex(mouseX: CGFloat, mouseY: CGFloat, iconCount: Int, screenWidth: CGFloat, screenHeight: CGFloat) -> Int? {
     let bgHeight = dockHeight + iconPadding * 2
     let bgY = screenHeight - bgHeight
     guard mouseY >= bgY && mouseY <= screenHeight else { return nil }
 
-    let totalBaseWidth = Float(iconCount) * baseIconSize + Float(iconCount - 1) * iconPadding
+    let totalBaseWidth = CGFloat(iconCount) * baseIconSize + CGFloat(iconCount - 1) * iconPadding
     let startX = (screenWidth - totalBaseWidth) / 2
 
     for i in 0..<iconCount {
-        let iconLeft = startX + Float(i) * (baseIconSize + iconPadding)
+        let iconLeft = startX + CGFloat(i) * (baseIconSize + iconPadding)
         if mouseX >= iconLeft && mouseX <= iconLeft + baseIconSize {
             return i
         }
@@ -70,9 +70,9 @@ func hoveredIconIndex(mouseX: Float, mouseY: Float, iconCount: Int, screenWidth:
 
 // MARK: - Declarative Dock View
 
-func dockView(state: DockState, width: Float, height: Float) -> some View {
+func dockView(state: DockState, width: CGFloat, height: CGFloat) -> some View {
     let sizes = magnifiedSizes(mouseX: state.mouseX, iconCount: items.count, screenWidth: width)
-    let totalWidth = sizes.reduce(0, +) + Float(items.count - 1) * iconPadding + iconPadding * 2
+    let totalWidth = sizes.reduce(0, +) + CGFloat(items.count - 1) * iconPadding + iconPadding * 2
     let bgHeight = dockHeight + iconPadding * 2
     let bgX = (width - totalWidth) / 2
     let bgY = height - bgHeight
@@ -108,7 +108,7 @@ func dockView(state: DockState, width: Float, height: Float) -> some View {
 
         // Minimized indicator dot
         if state.minimizedAppIds.contains(item.appId) {
-            let dotSize: Float = 4
+            let dotSize: CGFloat = 4
             let dotX = iconX + size / 2 - dotSize / 2
             let dotY = bgY + bgHeight - iconPadding / 2
             children.append(
@@ -121,8 +121,8 @@ func dockView(state: DockState, width: Float, height: Float) -> some View {
 
         // Hover label
         if i == hoveredIndex {
-            let textWidth = Float(item.name.count) * 8 + 20
-            let labelH: Float = 26
+            let textWidth = CGFloat(item.name.count) * 8 + 20
+            let labelH: CGFloat = 26
             let labelX = iconX + size / 2 - textWidth / 2
             let labelY = iconY - labelH - 6
 
@@ -163,12 +163,12 @@ struct DockApp: App {
         WindowConfiguration(title: "Dock", width: 1280, height: 800, role: .dock)
     }
 
-    func onPointerMove(x: Float, y: Float) {
+    func onPointerMove(x: CGFloat, y: CGFloat) {
         state.mouseX = x
         state.mouseY = y
     }
 
-    func onPointerButton(button: UInt32, pressed: Bool, x: Float, y: Float) {
+    func onPointerButton(button: UInt32, pressed: Bool, x: CGFloat, y: CGFloat) {
         if button == 0 && pressed {
             let index = hoveredIconIndex(mouseX: x, mouseY: y,
                                           iconCount: items.count,
